Move track related code to tracks.dart
