+++ packages/video_player_videohole/lib/src/video_player_tizen.dart
@@ -10,6 +10,7 @@ import 'package:flutter/widgets.dart';
 
 import '../video_player_platform_interface.dart';
 import 'messages.g.dart';
+import 'tracks.dart';
 
 /// An implementation of [VideoPlayerPlatform] that uses the
 /// Pigeon-generated [TizenVideoPlayerApi].
